requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langgraph>=0.0.55
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.7
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.5.0
jinja2>=3.1.2
aiofiles>=23.2.0
openai>=1.0.0

